refactored page order file by splitting entry instances into new file
