+++ planner/python/classes/page_entries/blank_entry.py
@@ -32,11 +32,11 @@ from classes.constants.addl_arg_keys import AddlArgKeys as Key
 from classes.elements.row_group import DualLineRowGroup
 from classes.elements.row_group import LineRowGroup
 
-from classes.page_layouts.half_letter_one_page import OnePageHalfLetter
+from classes.page_layouts.half_page_layout import HalfPageLayout
 
 
 #_______________________________________________________________________
-class BlankWrite(OnePageHalfLetter):
+class BlankWrite(HalfPageLayout):
   """
   Free write layout.
   """
